storage: move `SourceReader` pipeline into its own module
Pure code movement and fixing up use statements.
